requirements.txt
asgiref==3.4.1
autopep8==1.5.7
beautifulsoup4==4.9.3
Django==3.2.5
django-livereload-server==0.3.2
django-phonenumber-field==5.2.0
mysqlclient==2.0.3
phonenumberslite==8.12.29
Pillow==8.3.1
pycodestyle==2.7.0
pytz==2021.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
tornado==6.1
